Working Ring3!

+++ system/libc/unistd.c
@@ -4,15 +4,15 @@
 int syscall(int sysno, ...) {
     va_list args;
 
-    int arg0, arg1, arg2, arg3, arg4, arg5;
+    long int arg0, arg1, arg2, arg3, arg4, arg5;
 
     va_start(args, sysno);
-    arg0 = va_arg(args, int);
-    arg1 = va_arg(args, int);
-    arg2 = va_arg(args, int);
-    arg3 = va_arg(args, int);
-    arg4 = va_arg(args, int);
-    arg5 = va_arg(args, int);
+    arg0 = va_arg(args, long int);
+    arg1 = va_arg(args, long int);
+    arg2 = va_arg(args, long int);
+    arg3 = va_arg(args, long int);
+    arg4 = va_arg(args, long int);
+    arg5 = va_arg(args, long int);
     va_end(args);
 
     int retval;
